fix the bug of uint8_t and uint32_t compare in for loop

+++ Driver/Chip/AR8020/WirelessInfo/Inc/wireless_interface.h
@@ -227,7 +227,7 @@ void PAD_ENCODER_DYNAMIC_BITRATE_MODE_Handler(void *param);
 void PAD_ENCODER_DYNAMIC_BITRATE_SELECT_Handler(void *param);
 void PAD_WIRELESS_INFO_DISPLAY_Handler(void *param);
 void WIRELESS_ParseParamConfig(void *param);
-void convert_endian(void *data, uint8_t dataLen);
+void convert_endian(void *data, uint32_t dataLen);
 
 
 #endif
